Update Babydoge Bridge
